Make next gen connectors inherit cache and run methods
